New test case: app that collaboratively increase a counter
